requirements.txt
rich
pydantic
python-dotenv
requests
colorama
